hospital/requirements.txt
Django==5.2.7
gunicorn==23.0.0
python-decouple==3.8
python-dateutil==2.8.2
whitenoise==6.6.0
mozilla-django-oidc==4.0.1
google-cloud-kms
